+++ code/modules/power/cable_heavyduty.dm
@@ -12,45 +12,43 @@
 	color = null
 	var/static/allow_cutting = TRUE // Allows heavy cables to be cut by welder, up to server preference or admin vv during round. Changing it on one changes them all!
 
-/obj/structure/cable/heavyduty/attackby(obj/item/W, mob/user)
+/obj/structure/cable/heavyduty/wirecutter_act(mob/user, obj/item/W)
+	to_chat(user, span_notice("These cables are too tough to be cut with those [W.name]."))
+	return ITEM_INTERACT_BLOCKING
 
+/obj/structure/cable/heavyduty/welder_act(mob/user, obj/item/W)
 	var/turf/T = src.loc
-	if(!T.is_plating())
-		return
-
-	if(W.has_tool_quality(TOOL_WIRECUTTER))
-		to_chat(user, span_notice("These cables are too tough to be cut with those [W.name]."))
-		return
-	else if(W.has_tool_quality(TOOL_WELDER))
+	if(!T.is_plating() || !allow_cutting)
 		if(!allow_cutting)
 			to_chat(user, span_warning("Something in these cables make them too strong to cut!"))
-			return
+		return ITEM_INTERACT_BLOCKING
 
-		var/obj/item/weldingtool/WT = W.get_welder()
-		if(!WT.remove_fuel(2, user)) // Takes lots of fuel and time...
-			to_chat(user, span_infoplain("The welding tool must be on to complete this task."))
-			return
+	var/obj/item/weldingtool/WT = W.get_welder()
+	if(!WT.remove_fuel(2, user))
+		to_chat(user, span_infoplain("The welding tool must be on to complete this task."))
+		return ITEM_INTERACT_BLOCKING
 
-		playsound(src, WT.usesound, 50, 1)
-		if(do_after(user, 25 SECONDS * WT.toolspeed, target = src)) // Meant to be an obnoxiously long time due to these being intended as "mapper placed indestructable cables". However explosions can already break them if they turf change.
-			var/obj/item/stack/cable_coil/heavyduty/CC
-			if(src.d1)	// 0-X cables are 1 unit, X-X cables are 2 units long
-				CC = new/obj/item/stack/cable_coil/heavyduty(T, 2, color)
-			else
-				CC = new/obj/item/stack/cable_coil/heavyduty(T, 1, color)
+	playsound(src, WT.usesound, 50, 1)
+	if(do_after(user, 25 SECONDS * WT.toolspeed, target = src))
+		var/obj/item/stack/cable_coil/heavyduty/CC
+		if(src.d1)
+			CC = new/obj/item/stack/cable_coil/heavyduty(T, 2, color)
+		else
+			CC = new/obj/item/stack/cable_coil/heavyduty(T, 1, color)
 
-			src.add_fingerprint(user)
-			src.transfer_fingerprints_to(CC)
-			for(var/mob/O in viewers(src, null))
-				O.show_message(span_warning("[user] cuts the cable."), 1)
+		src.add_fingerprint(user)
+		src.transfer_fingerprints_to(CC)
+		for(var/mob/O in viewers(src, null))
+			O.show_message(span_warning("[user] cuts the cable."), 1)
 
-			qdel(src)
-		return
-	else if(istype(W, /obj/item/stack/cable_coil) && !istype(W, /obj/item/stack/cable_coil/heavyduty))
+		qdel(src)
+	return ITEM_INTERACT_SUCCESS
+
+/obj/structure/cable/heavyduty/attackby(obj/item/W, mob/user)
+	if(istype(W, /obj/item/stack/cable_coil) && !istype(W, /obj/item/stack/cable_coil/heavyduty))
 		to_chat(user, span_notice("You will need heavier cables to connect to these."))
 		return
-	else
-		..()
+	return ..()
 
 /obj/item/stack/cable_coil/heavyduty/turf_place(turf/simulated/F, mob/user)
 	if(istype(F, /turf/simulated/open))
